Dave 2020 9 exam 1 table D (#42)
* added some code for dynamic test

* added breaks between test

* fixed small bug in d2

* Allow caller to pass additional kwargs to utils.run_cmd

common.utils.run_cmd wraps subprocess.run and supplies defaults for some
kwargs. This commit allows the caller to pass additional arbitrary
kwargs to run_cmd to forward on to subprocess.run

This gives the caller more control over how the process is run. For
example, the caller can set a 5 second timeout by passing timeout=5

* Exam1 2020-09 - Use sed to delete the skel sleep in q2 & add timeout

* Delete the sleep call in the q2 main function in case students
accidentally left it in. Since sed matches this line using a fairly
specific pattern, it is unlikely to remove the line if the student
edited it (i.e. set their own sleep time). This is intentional.

* Add a timeout for the q2 dynamic analysis tests so that the grader
carries on in the face of race conditions that cause the program to hang
(i.e. num_dried++ is not protected by a mutex).

* Only delete sleep from skeleton code one time

It slipped my mind that we aren't checking out the code before each
dynamic test. Since we aren't, we only need to delete the sleep call
once.

* Bug fix: Call sed to delete skel sleep before compiling

>.>

* Update exam1 2020-09 rubric

* "Beautify" json

* Update rubric items and points assigned for D1 & D2

Co-authored-by: Evan Mesterhazy <evan.mesterhazy@gmail.com>